refactor: 주석 수정 및 컨트롤러 병합
