+++ server/src/main/java/com/oddok/server/domain/studyroom/dao/impl/HashtagQueryRepositoryImpl.java
@@ -4,13 +4,10 @@ import static com.oddok.server.domain.studyroom.entity.QHashtag.hashtag;
 import static com.oddok.server.domain.studyroom.entity.QStudyRoomHashtag.studyRoomHashtag;
 
 import com.oddok.server.domain.studyroom.dao.querydsl.HashtagQueryRepository;
-
-import com.oddok.server.domain.studyroom.entity.StudyRoom;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQuery;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import java.util.List;
-import java.util.Set;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
