+++ server/src/main/java/com/oddok/server/domain/studyroom/dao/impl/StudyRoomQueryRepository.java
@@ -14,7 +14,6 @@ import com.querydsl.jpa.impl.JPAQueryFactory;
 
 import java.time.LocalDate;
 import java.util.List;
-import java.util.Set;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
